Initial commit. A couple of scripts.
Added a ps script to check and remove folders redundancy for two dirs  parent/folders and parent/subfolder/folders. Keeping only the duplicated folders only in parent/folders.

Added a python script to check and print subfolders in a path which do not contain a .srt subtitle file. It is easy to modify to check for any other extension or substring in the filenames.